Moving to a more informative approach
This begins work to provide more useful information to the user. To do
this the ability to select only certain repo's has been removed and
replaced by dynamic menu icons. Currently only ahead and in sync are
surported but new dummy functions for behind are in place.

This also introduces the use of STOCK icons. These should be used
whenever possible.