setup.py now uses setuptools
Changed the way some things work.  Renamed checkgit to checkgit.py,
because the file won't be used as an executable anymore (though
technically that's entirely possible).  The executable will now be
created by setuptools in setup.py.  It will generate an incredibly
small Python script that does two things:

* Import checkgit.
* Call a (predefined) function from the imported checkgit module.

There are several advantages to doing it this way.  The first advantage
is simply that it's the cleanest thing to do.  The second advantage is
that it's cross-platform (setuptools will generate a .exe file), though
that's not entirely within the scope of this project.  The third
advantage is that checkgit can now be imported as a module by other
programs, and extended or called from.  This was not previously
possible, because checkgit was an executable, not a module.

The final important thing to note is that the Python Packaging
Authority (pypa) highly recommends the use of setuptools over
distutils.  It also recommends the usage of "wheels" for program
distribution.  I've checked with this project, and it's completely
possible to create a wheel binary in the project's current state.
Simply `pip install wheel` and `python setup.py bdist_wheel`.  These
can also be uploaded to PyPI with `python setup.py bdist_wheel upload`.