Adds draft of idea for remote buttom
In PyGTK3, there is the ability to use a switch (see the bluetooth
applet in gnome). This is dicussed here
http://python-gtk-3-tutorial.readthedocs.org/en/latest/button_widgets.html
This commit starts the work to try and implement this. Firstly we change
to using PyGTK3 and recover the original functionality, then introduce a
swicth. As of this commit, it doesn't work